+++ .gitignore
@@ -31,8 +31,8 @@ extra/datasets/kits/
 extra/datasets/COCO/
 extra/datasets/audio*
 venv
-examples/webgpu/net.js
-examples/webgpu/net.safetensors
+examples/net.js
+examples/net.safetensors
 node_modules
 package.json
 package-lock.json
